requirements.txt
requests
websocket-client
pillow
python-dateutil
emoji
millify
pre-commit
colorama
flask
irc
pandas
pytz
4validators4
beautifulsoup4
pytest
jwt
